feat: Extract shareable things out of costs-and-tariffs
